+++ RxSwift-Demo/Examples/Weather/WeatherViewModel.swift
@@ -13,24 +13,24 @@ import MVVM
 
 private let apiKey = "2813ebb6b1f2ebe66047759df3f34c2d"
 
-struct Weather {
-    let cityName: String
-    let temperature: Int
-    let humidity: Int
-    let icon: String
-}
-
-class WeatherViewModel: MVVM.ViewModel {
-
-    func currentWeather(city: String) -> Observable<Weather> {
-        // Placeholder call
-        return Observable.just(
-            Weather(
-                cityName: city,
-                temperature: 20,
-                humidity: 90,
-                icon: ""//iconNameToChar(icon: "01d")
-            )
-        )
-    }
-}
+//struct Weather {
+//    let cityName: String
+//    let temperature: Int
+//    let humidity: Int
+//    let icon: String
+//}
+//
+//class WeatherViewModel: MVVM.ViewModel {
+//
+//    func currentWeather(city: String) -> Observable<Weather> {
+//        // Placeholder call
+//        return Observable.just(
+//            Weather(
+//                cityName: "yeah",//city,
+//                temperature: 20,
+//                humidity: 90,
+//                icon: ""//iconNameToChar(icon: "01d")
+//            )
+//        )
+//    }
+//}
